+++ core/reflection/method.cpp
@@ -7,7 +7,7 @@ std::string Method::GetDescription(bool showOwnerType) const{
     ss << returnType.ToString() << " ";
     if (showOwnerType && ownerType) ss << ownerType->GetName() << "::";
     ss << name << "(";
-    for (int i = 0, n = paramTypes.size(); i<n; i++){
+    for (size_t i = 0, n = paramTypes.size(); i<n; i++){
         if (i != 0) ss << ", ";
         ss << paramTypes[i].ToString();
     }
