+++ core/reflection/method.h
@@ -24,7 +24,7 @@ protected:
     }
 
 public:
-    int GetParamCount() const{
+    size_t GetParamCount() const{
         return paramTypes.size();
     }
