+++ core/reflection/string_helper.h
@@ -8,7 +8,7 @@ public:
     static std::vector<std::string> Split(const std::string& str, char splitChar){
         std::vector<std::string> ret;
         int start = 0;
-        for (int i = 0, len = str.length(); i < len; i++){
+        for (size_t i = 0, len = str.length(); i < len; i++){
             if (str[i] == splitChar){
                 ret.push_back(str.substr(start, i - start));
                 start = i + 1;
@@ -20,7 +20,7 @@ public:
 
     static std::string RemoveSpaces(const std::string& str){
         std::vector<char> ret;
-        for (int i = 0, len = str.length(); i < len; i++){
+        for (size_t i = 0, len = str.length(); i < len; i++){
             char c = str[i];
             if (c != ' ' && c != '\t' && c != '\n'&& c != '\r')
                 ret.push_back(c);
